Fixing remarks for previous homework.
