+++ index.html
@@ -108,7 +108,7 @@
                     <textarea class="text__description" name="description" placeholder="Ваш комментарий..." 
                       required 
                       minlength="20" 
-                      maxlength = "140"/></textarea>
+                      maxlength="140"></textarea>
                 </fieldset>
 
                 <!-- Кнопка для отправки данных на сервер -->
